+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 
 # [1.0.0-beta.2](https://github.com/jhipster/generator-jhipster-nodejs/tree/v1.0.0-beta.2)
 
+- Testing entity relationship combinations with support for almost database field types [issue #81](https://github.com/jhipster/generator-jhipster-nodejs/issues/81)
+- Support blob types for the entity [issue #68](https://github.com/jhipster/generator-jhipster-nodejs/issues/68)
 - Disable testFramework and internationalization prompt questions with main app customization [issue #60](https://github.com/jhipster/generator-jhipster-nodejs/issues/60)
 - Provide test integration for quality in travis [issue #77](https://github.com/jhipster/generator-jhipster-nodejs/issues/77)